Add currency to expenses behaviour
