+++ commands.ini
@@ -1,9 +1,1 @@
 [Commands]
-camera:camera.take_photo
-video:camera.take_video
-goodbye|bye|we're going out|security on:security.on
-hello|we're (back|home|here)|security off:security.off
-test security:security.test
-is house empty:security.house_empty
-sweep:security.sweep
-override security:security.override
